Fix the way the main thread is determined.
syms_of_module is only called in an uninitialized Emacs before
dumping, so if we put the assignment of main_thread there, we will
always get the thread ID for the thread of the Emacs used for dumping,
not the current Emacs process.  Create a separate initialization
function instead that is unconditionally called from main.